uart: read chars until FIFO is empty
There might be more than one character in the FIFO when the interrupt is
finally processed. Change the semantics slightly to pull from the FIFO
until it is empty.

Also fix the interrupt race condition where the interrupt flag is
cleared after the FIFO was read. There is a small time window when a new
character arrives and we incorrectly clear the flag.